refactor: SocketMessage 공통부분(BaseMessage, MessageBody, AckBody 등등) application/dto/common 내부로 이동, MessageConverter(RequestMessageConverter, AckMessageJsonConverter) common/converter 패키지 내부로 이
