--- src/main/java/com/sera/chatting/application/dto/MessageBody.java
@@ -1,4 +1,0 @@
-package com.sera.chatting.application.dto;
-
-public interface MessageBody {
-}
