+++ src/main/java/com/sera/chatting/application/dto/common/MessageBody.java
@@ -1,0 +1,4 @@
+package com.sera.chatting.application.dto.common;
+
+public interface MessageBody {
+}
