+++ src/main/java/com/sera/chatting/application/dto/common/RequestBody.java
@@ -1,7 +1,9 @@
-package com.sera.chatting.application.dto;
+package com.sera.chatting.application.dto.common;
 
 import lombok.AllArgsConstructor;
+import lombok.Getter;
 
+@Getter
 @AllArgsConstructor
 public class RequestBody implements MessageBody {
 	private String command;
