+++ src/main/java/com/sera/chatting/common/converter/AckMessageJsonConverter.java
@@ -1,9 +1,9 @@
-package com.sera.chatting.application;
+package com.sera.chatting.common.converter;
 
 import org.springframework.stereotype.Component;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.sera.chatting.application.dto.AckMessage;
+import com.sera.chatting.application.dto.common.AckMessage;
 
 import lombok.RequiredArgsConstructor;
